+++ src/index.ts
@@ -2,6 +2,8 @@ import * as WUPHelpers from "./indexHelpers";
 
 export { WUPHelpers };
 
+export { default as WUPTimeObject } from "./externalTypes/timeObject";
+
 // export { default as WUPBaseElement } from "./baseElement";
 export { default as WUPPopupElement } from "./popup/popupElement";
 export { default as WUPSpinElement } from "./spinElement";